--- .github/workflows/batch-inference-guard.yml
@@ -1,102 +1,0 @@
-# Guards batched left-padded generation (issues #1066, #3699) against
-# regressions in _fast_prepare_inputs_for_generation and its wiring.
-#
-# History this protects against repeating:
-#   cc4c5d77 introduced position_ids = cache_position and a truncated
-#   attention mask (gibberish for padded rows, #1066); #2216 fixed the mask,
-#   #4100 fixed position_ids. Both bugs shipped because nothing tested this
-#   path. The two CPU tests below fail on either bug pattern.
-#
-# Fully deterministic on GPU-less hosted runners: layer 1 is stdlib-only AST
-# checks, layer 2 is a CPU-torch unit test through tests/conftest.py's CUDA
-# spoof. The matching staging proof lives on danielhanchen/unsloth-staging-2
-# (PR 170 green on clean code; PR 171 red within seconds at the AST step
-# after reintroducing the pre-#4100 bug).
-name: Batch inference guard
-
-on:
-  pull_request:
-    paths:
-      - 'unsloth/models/**'
-      - 'tests/utils/test_prepare_inputs*'
-      - '.github/workflows/batch-inference-guard.yml'
-  push:
-    branches: [main]
-    paths:
-      - 'unsloth/models/**'
-      - 'tests/utils/test_prepare_inputs*'
-      - '.github/workflows/batch-inference-guard.yml'
-  workflow_dispatch:
-
-permissions:
-  contents: read
-
-jobs:
-  leftpad-guard:
-    name: Left-padded generation guard (CPU)
-    runs-on: ubuntu-latest
-    timeout-minutes: 25
-    env:
-      PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION: python
-      UNSLOTH_COMPILE_DISABLE: '1'
-      UNSLOTH_IS_PRESENT: '1'
-    steps:
-      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd  # v6.0.2
-        with:
-          persist-credentials: false
-
-      - uses: actions/setup-python@a309ff8b426b58ec0e2a45f0f869d46889d02405  # v6.2.0
-        with:
-          python-version: '3.12'
-          cache: 'pip'
-
-      # Layer 1: structural AST guard. Needs only the Python stdlib + pytest,
-      # so it cannot be knocked out by dependency or import breakage.
-      - name: AST structural guard (no deps)
-        run: |
-          set -euxo pipefail
-          pip install pytest==9.0.3
-          python -m pytest tests/utils/test_prepare_inputs_ast_guard.py -v
-
-      # Layer 2: behavioral unit test against the real imported function,
-      # using the same CPU-runner dependency set as consolidated-tests-ci.yml
-      # (torch from the CPU wheel index, editable installs with --no-deps,
-      # GPU-less import handled by tests/conftest.py's CUDA spoof).
-      - name: Install CPU torch stack + unsloth
-        run: |
-          set -euxo pipefail
-          python -m pip install --upgrade pip
-          pip install \
-            'numpy<3' pytest==9.0.3 \
-            protobuf sentencepiece triton \
-            psutil packaging tqdm safetensors datasets \
-            'peft>=0.18,<0.20' 'accelerate>=0.34,<2'
-          pip install --index-url https://download.pytorch.org/whl/cpu \
-            'torch>=2.4,<2.11' 'torchvision<0.26'
-          pip install transformers trl
-          # bitsandbytes: hard import in unsloth/models/_utils.py; recent
-          # versions ship a CPU build that imports cleanly on Linux.
-          pip install 'bitsandbytes>=0.45'
-          pip install -e . --no-deps
-
-      - name: Clone and install unsloth_zoo (main, no deps)
-        run: |
-          set -euxo pipefail
-          for attempt in 1 2 3; do
-            rm -rf "$RUNNER_TEMP/unsloth-zoo"
-            if git clone --depth=1 https://github.com/unslothai/unsloth-zoo \
-                "$RUNNER_TEMP/unsloth-zoo"; then
-              break
-            fi
-            if [ "$attempt" -eq 3 ]; then
-              echo "::error::git clone unsloth-zoo failed after 3 attempts"
-              exit 1
-            fi
-            sleep $((5 * attempt))
-          done
-          pip install -e "$RUNNER_TEMP/unsloth-zoo" --no-deps
-
-      - name: Behavioral left-padding unit test
-        run: |
-          set -euxo pipefail
-          python -m pytest tests/utils/test_prepare_inputs_leftpad.py -v
